utils/db.py

# utils/db.py
from pymongo import MongoClient
import os
from dotenv import load_dotenv

load_dotenv()
mongo_uri = os.getenv("MONGODB_URI")
client = MongoClient(mongo_uri)
db = client['sentimentDB']
users = db['users']
results = db['results']