requirements.txt
Flask==2.1.1
pymongo[srv]==3.12.0
textblob==0.15.3
python-dotenv==1.0.1
gunicorn==20.1.0
